rename plugin files
Why:

* Follow file structure conventions
