+++ wazo_call_logd/plugins/api/plugin.py
@@ -1,7 +1,7 @@
 # Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resources import SwaggerResource
+from .http import SwaggerResource
 
 
 class Plugin:
