+++ wazo_call_logd/plugins/support_center/http.py
@@ -6,7 +6,7 @@ from xivo.auth_verifier import required_acl
 from xivo.tenant_flask_helpers import token, Tenant
 from wazo_call_logd.http import AuthResource
 
-from .schema import (
+from .schemas import (
     AgentStatisticsListRequestSchema,
     AgentStatisticsRequestSchema,
     AgentStatisticsSchemaList,
