+++ wazo_call_logd/plugins/support_center/plugin.py
@@ -1,14 +1,14 @@
 # Copyright 2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resource import (
+from .http import (
     AgentsStatisticsResource,
     AgentStatisticsResource,
     QueuesStatisticsResource,
     QueueStatisticsResource,
 )
 
-from .service import AgentStatisticsService, QueueStatisticsService
+from .services import AgentStatisticsService, QueueStatisticsService
 
 
 class Plugin(object):
